requirements.txt
# Core Telegram Bot
python-telegram-bot>=20.0
python-dotenv

# Web App Server
flask
flask-cors

# AI Providers (optional - at least one required)
google-generativeai
openai

# Utilities
requests
